Integration tests: delete back button test
The test seems too flaky nowadays -> practically it only fails in Cypress, while it always works with real user. 

It also fails unpredictably: these tests are run in different Node versions & different OSs. It usually passes in _most_ of them, but fails in one. And the one it fails in wont remain the same.

Just nuke the test, not worth the trouble.